update: serverプログラムの型安全化、及びリファクタリング。サーバークライアントともにビルドは通る。

+++ modules/API/ClockClient.cpp
@@ -2,7 +2,7 @@
 
 ClockClient::ClockClient(SpikeClient& client) : spikeClient(client) {}
 
-void ClockClient::sleep(uint32_t milliseconds) {
+void ClockClient::sleep(uint64_t milliseconds) {
     spike::ClockSleepRequest req{ milliseconds };
     spikeClient.executeCommand(
         spike::CommandId::CLOCK_SLEEP,
@@ -10,8 +10,8 @@ void ClockClient::sleep(uint32_t milliseconds) {
     );
 }
 
-uint32_t ClockClient::now() {
-    auto res = spikeClient.executeQuery<uint32_t>(
+uint64_t ClockClient::now() {
+    auto res = spikeClient.executeQuery<uint64_t>(
         spike::CommandId::CLOCK_NOW
     );
     return res.value_or(0);
